requirements.txt
Django
djangorestframework
markdown
django-filter
django-cors-headers
uwsgi